formatting and checks
